LibCrypto: Move hash constructors out of line
