+++ Libraries/LibCrypto/Hash/MD5.h
@@ -15,10 +15,7 @@ class MD5 final : public OpenSSLHashFunction<MD5, 512, 128> {
     AK_MAKE_NONCOPYABLE(MD5);
 
 public:
-    explicit MD5(EVP_MD_CTX* context)
-        : OpenSSLHashFunction(EVP_md5(), context)
-    {
-    }
+    explicit MD5(EVP_MD_CTX* context);
 
     virtual ByteString class_name() const override
     {
