+++ Libraries/LibCrypto/Hash/SHA2.h
@@ -15,10 +15,7 @@ class SHA256 final : public OpenSSLHashFunction<SHA256, 512, 256> {
     AK_MAKE_NONCOPYABLE(SHA256);
 
 public:
-    explicit SHA256(EVP_MD_CTX* context)
-        : OpenSSLHashFunction(EVP_sha256(), context)
-    {
-    }
+    explicit SHA256(EVP_MD_CTX* context);
 
     virtual ByteString class_name() const override
     {
@@ -30,10 +27,7 @@ class SHA384 final : public OpenSSLHashFunction<SHA384, 1024, 384> {
     AK_MAKE_NONCOPYABLE(SHA384);
 
 public:
-    explicit SHA384(EVP_MD_CTX* context)
-        : OpenSSLHashFunction(EVP_sha384(), context)
-    {
-    }
+    explicit SHA384(EVP_MD_CTX* context);
 
     virtual ByteString class_name() const override
     {
@@ -45,10 +39,7 @@ class SHA512 final : public OpenSSLHashFunction<SHA512, 1024, 512> {
     AK_MAKE_NONCOPYABLE(SHA512);
 
 public:
-    explicit SHA512(EVP_MD_CTX* context)
-        : OpenSSLHashFunction(EVP_sha512(), context)
-    {
-    }
+    explicit SHA512(EVP_MD_CTX* context);
 
     virtual ByteString class_name() const override
     {
